stripped MetaDataKeeper from all extraneous stuff...
