+++ src/main/java/org/corfudb/infrastructure/wireprotocol/NettyCollectResponseMsg.java
@@ -1,32 +1,33 @@
 package org.corfudb.infrastructure.wireprotocol;
 
-import com.sun.corba.se.impl.orbutil.ObjectWriter;
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.ByteBufInputStream;
 import io.netty.buffer.ByteBufOutputStream;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import javax.json.*;
-import java.io.*;
-import java.util.HashSet;
-import java.util.Set;
-import java.util.UUID;
+import javax.json.Json;
+import javax.json.JsonObject;
+import javax.json.JsonReader;
+import javax.json.JsonWriter;
+import java.io.BufferedReader;
+import java.io.StringReader;
 
 /**
  * Created by dmalkhi on 11/9/15.
  */
+@Getter
+@Setter
 @NoArgsConstructor
-public class NettyLayoutServerRequestMsg extends NettyCorfuMsg {
+public class NettyCollectResponseMsg extends NettyCorfuMsg {
 
     JsonObject jo = null;
 
-    public NettyLayoutServerRequestMsg(JsonObject jsonObject)
+    public NettyCollectResponseMsg(JsonObject jo)
     {
-        this.msgType = NettyCorfuMsgType.LAYOUT_REQ;
-        this.jo = jsonObject;
+        this.jo = jo;
+        this.msgType = NettyCorfuMsgType.META_COLLECT_RES;
     }
 
     /**
